refactor: rename workflow helper headers
