+++ include/rarexsec/flow/SnapshotBuilder.h
