+++ src/run/study_all.cpp
@@ -1,8 +1,8 @@
 #include <rarexsec/flow/Study.h>
-#include <rarexsec/flow/Helpers.h>
-#include <rarexsec/flow/Display.h>
-#include <rarexsec/flow/Plots.h>
-#include <rarexsec/flow/Snapshot.h>
+#include <rarexsec/flow/Where.h>
+#include <rarexsec/flow/EventDisplayBuilder.h>
+#include <rarexsec/flow/PlotBuilders.h>
+#include <rarexsec/flow/SnapshotBuilder.h>
 using namespace analysis::dsl;
 
 int main() {
